feat: 승인여부에 따라 삭제가능, 불가능 설정

+++ backend/src/main/java/com/ssafy/ourdoc/domain/bookreport/service/BookReportStudentService.java
@@ -70,6 +70,10 @@ public class BookReportStudentService {
 		BookReport bookReport = bookReportRepository.findById(bookReportId)
 			.orElseThrow(() -> new NoSuchElementException("지울 독서록이 없습니다."));
 
+		if (bookReport.getApproveTime() == null) {
+			throw new IllegalArgumentException("승인이 된 독서록은 지울 수 없습니다.");
+		}
+
 		bookReportRepository.delete(bookReport);
 	}
